requirements.txt
plyfile
opencv-python
trimesh==2.35.39
tensorboardX
easydict
tqdm
h5py
matplotlib
pyyaml
imageio
pandas